+++ ssvm/ssvm.py
@@ -33,8 +33,10 @@ import time
 
 from collections import OrderedDict
 from typing import List, Tuple, Dict, Union, Optional, Callable
+
 from sklearn.utils.validation import check_random_state
 from sklearn.metrics import ndcg_score
+
 from joblib.parallel import Parallel, delayed
 from joblib.memory import Memory
 
@@ -44,13 +46,13 @@ from msmsrt_scorer.lib.cindex_measure import cindex
 
 import ssvm.cfg
 from ssvm.data_structures import SequenceSample, Sequence, LabeledSequence, SpanningTrees
-from ssvm.data_structures import ABCCandSQLiteDB
 from ssvm.factor_graphs import get_random_spanning_tree
-from ssvm.kernel_utils import _min_max_dense_jit, generalized_tanimoto_kernel_FAST
+from ssvm.kernel_utils import _min_max_dense_jit, generalized_tanimoto_kernel_FAST, rbf_kernel
 from ssvm.kernel_utils import _min_max_dense_ufunc, _min_max_dense_ufunc_int, tanimoto_kernel_FAST
 from ssvm.utils import item_2_idc
 from ssvm.dual_variables import DualVariables
 from ssvm.ssvm_meta import _StructuredSVM
+from ssvm.loss_functions import hamming_loss, tanimoto_loss, minmax_loss, generalized_tanimoto_loss, kernel_loss
 
 
 JOBLIB_CACHE = Memory(ssvm.cfg.JOBLIB_MEMORY_CACHE_LOCATION, verbose=0)
@@ -76,24 +78,44 @@ class StructuredSVMSequencesFixedMS2(_StructuredSVM):
     """
     Structured Support Vector Machine (SSVM) for (MS, RT)-sequence classification.
     """
-    def __init__(self, mol_feat_label_loss: str, mol_feat_retention_order: str,
-                 mol_kernel: Union[str, Callable[[np.ndarray, np.ndarray], np.ndarray]], n_trees_per_sequence: int = 1,
-                 n_jobs: int = 1, update_direction: str = "map", *args, **kwargs):
-        """
-        :param mol_feat_label_loss:
-        :param mol_feat_retention_order:
-        :param mol_kernel:
-        :param n_trees_per_sequence: scalar, number of spanning trees per sequence.
-        :param args:
-        :param kwargs:
+    def __init__(
+            self, mol_feat_label_loss: str, mol_feat_retention_order: str,
+            mol_kernel: Union[str, Callable[[np.ndarray, np.ndarray], np.ndarray]], n_trees_per_sequence: int = 1,
+            n_jobs: int = 1, update_direction: str = "map", gamma: Optional[float] = None, label_loss: str = "kernel_loss",
+            *args, **kwargs
+    ):
         """
 
+        """
         self.mol_feat_label_loss = mol_feat_label_loss
         self.mol_feat_retention_order = mol_feat_retention_order
-        self.mol_kernel = self.get_mol_kernel(mol_kernel)
+        self.mol_kernel = self.get_mol_kernel(mol_kernel, kernel_parameters={"gamma": gamma})
         self.n_trees_per_sequence = n_trees_per_sequence
         self.n_jobs = n_jobs
         self.update_direction = update_direction
+        self.label_loss = label_loss
+
+        if self.label_loss == "hamming":
+            raise NotImplementedError(
+                "Currently the Hamming-loss implementation is not really a hamming-loss on the label sequences."
+            )
+        elif self.label_loss == "tanimoto_loss":
+            self.label_loss_fun = tanimoto_loss
+        elif self.label_loss == "minmax_loss":
+            self.label_loss_fun = minmax_loss
+        elif self.label_loss == "generalized_tanimoto_loss":
+            self.label_loss_fun = generalized_tanimoto_loss
+        elif self.label_loss == "kernel_loss":
+            # Generalization of the kernel losses, the kernel functions for the features is used for the loss as well.
+            def _label_loss(y, Y):
+                return kernel_loss(y, Y, self.mol_kernel)
+
+            self.label_loss_fun = _label_loss
+        else:
+            raise ValueError(
+                "Invalid label loss '%s'. Choices are 'hamming', 'tanimoto_loss', 'minmax_loss', "
+                "'generalized_tanimoto_loss' and 'kernel_loss'."
+            )
 
         if self.n_trees_per_sequence > 1:
             raise ValueError("Currently only a single spanning tree per sequence is supported.")
@@ -1192,12 +1214,19 @@ class StructuredSVMSequencesFixedMS2(_StructuredSVM):
                     tf.summary.scalar(k, data=v, step=n_iter)
 
     @staticmethod
-    def get_mol_kernel(mol_kernel: Union[str, Callable[[np.ndarray, np.ndarray], np.ndarray]]) \
-            -> Callable[[np.ndarray, np.ndarray], np.ndarray]:
+    def get_mol_kernel(
+            mol_kernel: Union[str, Callable[[np.ndarray, np.ndarray], np.ndarray]],
+            kernel_parameters: Optional[Dict] = None
+    ) -> Callable[[np.ndarray, np.ndarray], np.ndarray]:
         """
 
-        :param mol_kernel:
-        :return:
+        :param mol_kernel: string or callable
+            string: name of the kernel function
+            callable: kernel function
+
+        :param kernel_parameters: dictionary, containing the kernel parameters for the specified kernel
+
+        :return: callable, kernel function
         """
         if callable(mol_kernel):
             return mol_kernel
@@ -1211,6 +1240,12 @@ class StructuredSVMSequencesFixedMS2(_StructuredSVM):
             return _min_max_dense_ufunc
         elif mol_kernel == "minmax_ufunc_int":
             return _min_max_dense_ufunc_int
+        elif mol_kernel == "rbf":
+            # Define wrapper around the RBF kernel to bind the specified gamma parameter value
+            def _rbf_kernel(X, Y):
+                return rbf_kernel(X, Y, gamma=kernel_parameters["gamma"])
+
+            return _rbf_kernel
         else:
             raise ValueError("Invalid molecule kernel: '{}'".format(mol_kernel))